Moved setUp/teardown class to WebTestBase
